bonus pity menu WIP

+++ paper/src/main/java/com/badbones69/crazycrates/api/PrizeManager.java
@@ -181,7 +181,7 @@ public class PrizeManager {
                 for (Tier tier : crate.getTiers()) {
                     double chance = tier.getChance();
 
-                    int num = MiscUtils.useOtherRandom() ? ThreadLocalRandom.current().nextInt(tier.getMaxRange()) : new Random().nextInt(tier.getMaxRange());
+                    double num = MiscUtils.useOtherRandom() ? ThreadLocalRandom.current().nextDouble(tier.getMaxRange()) : new Random().nextDouble(tier.getMaxRange());
 
                     if (num >= 1 && num <= chance) {
                         return tier;
